requirements.txt
requests~=2.25.1
pip~=23.1.2
distro~=1.7.0
cryptography~=3.4.8
Pillow~=9.5.0
tornado~=6.4
Jinja2~=3.1.2
keyring~=23.5.0
click~=8.0.3
ipython~=8.18.1
certifi~=2020.6.20
chardet~=4.0.0
idna~=2.10
urllib3~=1.26.5
setuptools~=56.0.0
packaging~=23.2
pyparsing~=2.4.7
zipp~=3.17.0
future~=0.18.2
python-dotenv~=1.0.0